requirements.txt
Django~=5.2.10
dotenv~=0.9.9
python-dotenv~=1.2.1
requests~=2.32.5
pdfplumber
google-generativeai